Added eventual public API for Parcel crate, moved napi binding code to node_bindings crate
