+++ crates/node-bindings/src/core/mod.rs
@@ -1,2 +1,4 @@
 //! JavaScript API for `parcel_core::requests`
 pub mod js_requests;
+pub mod project_path;
+pub mod requests;
