+++ crates/node-bindings/src/core/project_path.rs
