+++ crates/parcel_core/src/lib.rs
@@ -1,3 +1,4 @@
+#![deny(unused_crate_dependencies)]
 //! Core re-implementation in Rust
 
 pub mod bundle_graph;
@@ -5,8 +6,4 @@ pub mod hash;
 pub mod plugin;
 pub mod types;
 
-/// New-type for paths relative to a project-root
-pub mod project_path;
-
-/// Request types and run functions
-pub mod requests;
+// pub use parcel::*;
