+++ crates/parcel_core/src/parcel/mod.rs
@@ -1,0 +1,2 @@
+mod parcel;
+pub use parcel::*;
