+++ crates/parcel_core/src/parcel/parcel.rs
@@ -1,0 +1,12 @@
+// TODO eventual public API for Parcel
+pub struct Parcel {}
+
+impl Parcel {
+  pub fn run() {
+    todo!();
+  }
+
+  pub fn watch() {
+    todo!();
+  }
+}
